move cache code to cache folder
